refactor: split explorer sidebar controls
